This fixes an integer overflow issue creating negative results, Issue https://github.com/c-a-m/passfault/issues/8

+++ core/src/com/passfault/dictionary/CandidatePattern.java
@@ -78,8 +78,8 @@ public class CandidatePattern {
     return clone;
   }
 
-  public PasswordPattern getMatchingPattern(int dictionarySize, String strategyName, String classification, int currentPosition) {
-    int crackSize = dictionarySize;
+  public PasswordPattern getMatchingPattern(double dictionarySize, String strategyName, String classification, int currentPosition) {
+    double crackSize = dictionarySize;
     String description = "Match";
     if (decorator != null) {
       description = decorator.getDescription();
